Touch Events
There's a handleTouch function, seperate from handleMouse, mostly for the preventDefault call to prevent extra click events.
Which chrome doesn't like but it seems to work anyway - looks like we need to add { passive: false } to the event listener to get this working properly

The mouseIntersect function is largely the same as the previous intersect function.
I've pulled out the raycasting function mostly so it can be used multiple times for every touch event
The touchIntersect function also adds the raycast targets for the other touch points - I'm not sure it's needed but I have a feeling it will be useful.

I changed the data structure a bit, feel free to change it back. I just felt it was a little messy mixing up the three obj with the event data.