complete the fix for attaching to non-Object3D
When I added the extra checks for parentInstance.isObject3D, I only added them when mounting the object and missed out the case in removeChild. It seemed to work when I tested against my application, but now removeChild throws an error  in my application, trying to call remove on something that is not Object3D. This change fixes that, making the conditions for calling add and remove the same.