fix unmounting pointer capture target
Previously, unmounting the pointer capture target would not release its pointer capture. This would obviously cause problems when sending subsequent pointer events to it. This commit fixes the problem by storing the native (DOM) pointer capture target in the capturesMap so that it can be removed in removeInteractivity.

Adding a new test for pointer capture necessitated changing how the event props are copied in handleIntersects. The old code looked at the prototype (which should be PointerEvent) to find the props to copy, but in tests the pointerEventPolyfill is active and the prototype is MouseEvent, which meant none of the PointerEvent properties got copied. I've changed it to use for...in to look at the object's own properties and its inherited properties.